create solar system component and refactor
